add: 목록에서 운동 클릭 시, add페이지로 이동 및 담긴 값 계산 구현 -추가 구현할 내용: 목록에서 운동 클릭 시, 쿠키에 값 저장 -고민중인 내용: 입력들어오는 값 테이블을 만들지, div로 진행할지,,고민

+++ src/main/java/xyz/teamcarrot/myct/healthInfo/HealthDicService.java
@@ -4,5 +4,5 @@ import java.util.List;
 import java.util.Map;
 
 public interface HealthDicService {
-	List<Map<String, Object>> getHealthDic(String healthName, int kg, int minute);
+	List<Map<String, Object>> getHealthDic(String healthName, double kg, int minute);
 }
